renamed file to boot.rb from application_config.rb
