+++ lib/publisher.rb
@@ -1,4 +1,4 @@
-require File.join(Application.root, 'application_config')
+require File.join(Application.root, 'boot')
 
 class Publisher
